# apps/server/src/routes/auth.ts
/**
 * Authentication Routes with OpenAPI Documentation
 */

import { Router } from "express";
import { createRateLimit, RateLimitPresets } from "../middleware/rate-limiter";
import { ValidationPresets } from "../middleware/validation";

const router = Router();

/**
 * @swagger
 * /api/auth/signup:
 *   post:
 *     tags: [Authentication]
 *     summary: Register a new user account
 *     description: Create a new user account with email and password
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             $ref: '#/components/schemas/UserCreateRequest'
 *           example:
 *             email: "user@example.com"
 *             password: "SecurePass123!"
 *             firstName: "John"
 *             lastName: "Doe"
 *             organizationName: "Acme Corp"
 *             acceptTerms: true
 *     responses:
 *       201:
 *         description: User account created successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                   example: true
 *                 message:
 *                   type: string
 *                   example: "Account created successfully"
 *                 user:
 *                   $ref: '#/components/schemas/User'
 *             example:
 *               success: true
 *               message: "Account created successfully"
 *               user:
 *                 id: "123e4567-e89b-12d3-a456-426614174000"
 *                 email: "user@example.com"
 *                 firstName: "John"
 *                 lastName: "Doe"
 *                 role: "ADMIN"
 *                 createdAt: "2023-01-01T00:00:00Z"
 *       400:
 *         $ref: '#/components/responses/ValidationError'
 *       409:
 *         description: Email already exists
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Error'
 *             example:
 *               code: "EMAIL_EXISTS"
 *               message: "An account with this email already exists"
 *       429:
 *         $ref: '#/components/responses/RateLimitError'
 *     security: []
 */
router.post(
	"/signup",
	createRateLimit(RateLimitPresets.auth),
	ValidationPresets.auth.signUp,
	async (req, res, next) => {
		try {
			// Implementation would go here
			res.status(201).json({
				success: true,
				message: "Account created successfully",
				user: {
					id: "123e4567-e89b-12d3-a456-426614174000",
					email: req.body.email,
					firstName: req.body.firstName,
					lastName: req.body.lastName,
					role: "ADMIN",
					createdAt: new Date().toISOString(),
				},
			});
		} catch (error) {
			next(error);
		}
	},
);

/**
 * @swagger
 * /api/auth/signin:
 *   post:
 *     tags: [Authentication]
 *     summary: Sign in to user account
 *     description: Authenticate user with email and password
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email, password]
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *                 description: User email address
 *               password:
 *                 type: string
 *                 description: User password
 *           example:
 *             email: "user@example.com"
 *             password: "SecurePass123!"
 *     responses:
 *       200:
 *         description: Authentication successful
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                 message:
 *                   type: string
 *                 accessToken:
 *                   type: string
 *                   description: JWT access token
 *                 refreshToken:
 *                   type: string
 *                   description: Refresh token for obtaining new access tokens
 *                 expiresIn:
 *                   type: integer
 *                   description: Token expiration time in seconds
 *                 user:
 *                   $ref: '#/components/schemas/User'
 *             example:
 *               success: true
 *               message: "Authentication successful"
 *               accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
 *               refreshToken: "refresh_token_here"
 *               expiresIn: 3600
 *               user:
 *                 id: "123e4567-e89b-12d3-a456-426614174000"
 *                 email: "user@example.com"
 *                 firstName: "John"
 *                 lastName: "Doe"
 *       400:
 *         $ref: '#/components/responses/ValidationError'
 *       401:
 *         description: Invalid credentials
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Error'
 *             example:
 *               code: "INVALID_CREDENTIALS"
 *               message: "Invalid email or password"
 *       429:
 *         $ref: '#/components/responses/RateLimitError'
 *     security: []
 */
router.post(
	"/signin",
	createRateLimit(RateLimitPresets.auth),
	ValidationPresets.auth.signIn,
	async (req, res, next) => {
		try {
			// Implementation would go here
			res.json({
				success: true,
				message: "Authentication successful",
				accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
				refreshToken: "refresh_token_here",
				expiresIn: 3600,
				user: {
					id: "123e4567-e89b-12d3-a456-426614174000",
					email: req.body.email,
					firstName: "John",
					lastName: "Doe",
				},
			});
		} catch (error) {
			next(error);
		}
	},
);

/**
 * @swagger
 * /api/auth/refresh:
 *   post:
 *     tags: [Authentication]
 *     summary: Refresh access token
 *     description: Obtain a new access token using a refresh token
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [refreshToken]
 *             properties:
 *               refreshToken:
 *                 type: string
 *                 description: Valid refresh token
 *           example:
 *             refreshToken: "refresh_token_here"
 *     responses:
 *       200:
 *         description: Token refreshed successfully
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                 accessToken:
 *                   type: string
 *                 expiresIn:
 *                   type: integer
 *             example:
 *               success: true
 *               accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
 *               expiresIn: 3600
 *       400:
 *         $ref: '#/components/responses/ValidationError'
 *       401:
 *         description: Invalid refresh token
 *         content:
 *           application/json:
 *             schema:
 *               $ref: '#/components/schemas/Error'
 *             example:
 *               code: "INVALID_REFRESH_TOKEN"
 *               message: "Refresh token is invalid or expired"
 *     security: []
 */
router.post(
	"/refresh",
	ValidationPresets.auth.refreshToken,
	async (_req, res, next) => {
		try {
			// Implementation would go here
			res.json({
				success: true,
				accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
				expiresIn: 3600,
			});
		} catch (error) {
			next(error);
		}
	},
);

/**
 * @swagger
 * /api/auth/logout:
 *   post:
 *     tags: [Authentication]
 *     summary: Sign out user
 *     description: Invalidate user session and tokens
 *     responses:
 *       200:
 *         description: Logout successful
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                 message:
 *                   type: string
 *             example:
 *               success: true
 *               message: "Logout successful"
 *       401:
 *         $ref: '#/components/responses/UnauthorizedError'
 */
router.post("/logout", async (_req, res, next) => {
	try {
		// Implementation would go here
		res.json({
			success: true,
			message: "Logout successful",
		});
	} catch (error) {
		next(error);
	}
});

/**
 * @swagger
 * /api/auth/password/reset:
 *   post:
 *     tags: [Authentication]
 *     summary: Request password reset
 *     description: Send password reset email to user
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required: [email]
 *             properties:
 *               email:
 *                 type: string
 *                 format: email
 *                 description: User email address
 *           example:
 *             email: "user@example.com"
 *     responses:
 *       200:
 *         description: Password reset email sent
 *         content:
 *           application/json:
 *             schema:
 *               type: object
 *               properties:
 *                 success:
 *                   type: boolean
 *                 message:
 *                   type: string
 *             example:
 *               success: true
 *               message: "If an account with this email exists, a password reset link has been sent"
 *       400:
 *         $ref: '#/components/responses/ValidationError'
 *       429:
 *         $ref: '#/components/responses/RateLimitError'
 *     security: []
 */
router.post(
	"/password/reset",
	createRateLimit(RateLimitPresets.auth),
	ValidationPresets.auth.resetPassword,
	async (_req, res, next) => {
		try {
			// Implementation would go here
			res.json({
				success: true,
				message:
					"If an account with this email exists, a password reset link has been sent",
			});
		} catch (error) {
			next(error);
		}
	},
);

export default router;